requirements.txt
flask>=3.0
google-auth>=2.29
google-auth-oauthlib>=1.2
google-api-python-client>=2.125
python-dotenv>=1.0
gunicorn>=21.2
openpyxl>=3.1
tzdata>=2024.1
